+++ crates/pytauri-core/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+## [0.5.0]
+
 ### BREAKING
 
 - [#141](https://github.com/pytauri/pytauri/pull/141) - feat!: `pytauri.path.PathResolver` now returns a `pathlib.Path` object instead of a `str`.
@@ -88,6 +90,7 @@
 ## [0.1.0-beta.0]
 
 [unreleased]: https://github.com/pytauri/pytauri/tree/HEAD
+[0.5.0]: https://github.com/pytauri/pytauri/releases/tag/rs/pytauri-core/v0.5.0
 [0.4.0]: https://github.com/pytauri/pytauri/releases/tag/rs/pytauri-core/v0.4.0
 [0.3.0]: https://github.com/pytauri/pytauri/releases/tag/rs/pytauri-core/v0.3.0
 [0.2.0]: https://github.com/pytauri/pytauri/releases/tag/rs/pytauri-core/v0.2.0